requirements.txt
Flask
tensorflow
pillow
numpy
gdown
gunicorn